XBEE Module
Zigbee is a popular wireless communication protocol used to transfer a small amount of data with very low power. It is widely used in applications where data has to be shared among many nodes within personal space and with the advent of the Internet of things (IoT). Let’s take that a step further and learn how to establish Zigbee communication using Arduino and NodeMCU boards.